+++ vite.config.ts
@@ -33,12 +33,12 @@ export default defineConfig({
   },
   server: {
     host: "127.0.0.1",
-    port: 5173,
+    port: 5193,
     strictPort: true
   },
   preview: {
     host: "127.0.0.1",
-    port: 4173,
+    port: 4193,
     strictPort: true
   }
 });